temp_extract/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
aiofiles>=23.2.1
msgpack>=1.0.7
prometheus-client>=0.19.0
structlog>=24.1.0